requirements.txt
langchain
langchain-openai
langchain-community
supabase
python-frontmatter